TSC 312 added text and fixed alignment issues of buttons to shift rows
