+++ app/src/settings_tabs/ColumnMenu.css
@@ -30,3 +30,32 @@
   margin-top: 10px;
   margin-left: 10px;
 }
+
+.accordion-position-controls {
+  display: flex;
+  align-items: center;
+  margin-top: 10px;
+  margin-left: 10px;
+  margin-bottom: 10px;
+}
+
+.lightgray-square {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 35px;
+  height: 35px;
+  background-color: lightgray;
+  margin: 0 5px;
+  border-radius: 5px;
+}
+
+.custom-icon-button {
+  background-color: transparent;
+  border: none;
+  padding-left: 17px;
+  margin: 0;
+  width: 100%;
+  height: 100%;
+  border-radius: 5px; /* Match the border-radius of the parent */
+}
